+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improved the permission selector with icons in the create or update access dialog
+- Extracted the access level icon to a reusable component
 - Disabled the telemetry in the _Storybook_ setup
 - Upgraded the `Node.js` engine from version `>=22.18.0` to `>=22.22.3` (`package.json`)